requirements.txt
gradio
prompt-toolkit
requests==2.32.5
python-dotenv==1.1.1
